feat(saas): 通知/登录日志枚举契约对齐 + 通知中心与收藏夹样式统一
- UserInboxItemDto 通知类型/状态、ProfileLoginLogDto 登录结果由 int 改为枚举（随全局 JsonStringEnumConverter 输出字符串），修复铃铛未读角标不显示与登录日志「状态10」展示
- 前端 LoginLogItem.loginResult 改为字符串枚举类型，登录日志标签/配色/图标按审计事件语义映射；条目紧凑两行布局、每页 10 条一屏可览
- 通知中心弹层收窄至 420px，未读角标增加 processing 脉冲动画，头部/底部样式收紧
- 收藏夹面板头部与表格设置统一（text-base 标题 + 同步徽标 + 分隔线），操作说明移至底部灰字，保底高度避免空态过矮

+++ backend/src/modules/XiHan.BasicApp.Saas/Application/QueryServices/Messaging/UserInboxQueryService.cs
@@ -89,7 +89,7 @@ public sealed class UserInboxQueryService
         return [.. userNotifications
             .Where(item => notificationMap.ContainsKey(item.NotificationId))
             .Select(item => UserNotificationDispatchService.ToInboxItem(item, notificationMap[item.NotificationId]))
-            .Where(item => !unreadOnly || item.NotificationStatus == (int)NotificationStatus.Unread || (item.NeedConfirm && !item.ConfirmTime.HasValue))
+            .Where(item => !unreadOnly || item.NotificationStatus == NotificationStatus.Unread || (item.NeedConfirm && !item.ConfirmTime.HasValue))
             .OrderByDescending(item => item.SendTime)
             .ThenByDescending(item => item.BasicId)
             .Take(MaxInboxItems)];
